third_party/blink/web_tests/platform/generic/fast/forms/input-type-change3-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.type is "text"
PASS input.getAttribute('type') is null
PASS input.type for "text" is correctly "text".
PASS input.getAttribute("type") for "text" is correctly "text".
PASS input.type for "TEXT" is correctly "text".
PASS input.getAttribute("type") for "TEXT" is correctly "TEXT".
PASS input.type for " text " is correctly "text".
PASS input.getAttribute("type") for " text " is correctly " text ".
PASS input.type for "button" is correctly "button".
PASS input.getAttribute("type") for "button" is correctly "button".
PASS input.type for " button " is correctly "text".
PASS input.getAttribute("type") for " button " is correctly " button ".
PASS input.type for "checkbox" is correctly "checkbox".
PASS input.getAttribute("type") for "checkbox" is correctly "checkbox".
PASS input.type for "checKbox" is correctly "text".
PASS input.getAttribute("type") for "checKbox" is correctly "checKbox".
PASS input.type for "email" is correctly "email".
PASS input.getAttribute("type") for "email" is correctly "email".
PASS input.type for "file" is correctly "file".
PASS input.getAttribute("type") for "file" is correctly "file".
PASS input.type for "hidden" is correctly "hidden".
PASS input.getAttribute("type") for "hidden" is correctly "hidden".
PASS input.type for "image" is correctly "image".
PASS input.getAttribute("type") for "image" is correctly "image".
PASS input.type for "isindex" is correctly "text".
PASS input.getAttribute("type") for "isindex" is correctly "isindex".
PASS input.type for "number" is correctly "number".
PASS input.getAttribute("type") for "number" is correctly "number".
PASS input.type for "password" is correctly "password".
PASS input.getAttribute("type") for "password" is correctly "password".
PASS input.type for "passwd" is correctly "text".
PASS input.getAttribute("type") for "passwd" is correctly "passwd".
PASS input.type for "radio" is correctly "radio".
PASS input.getAttribute("type") for "radio" is correctly "radio".
PASS input.type for "range" is correctly "range".
PASS input.getAttribute("type") for "range" is correctly "range".
PASS input.type for "reset" is correctly "reset".
PASS input.getAttribute("type") for "reset" is correctly "reset".
PASS input.type for "search" is correctly "search".
PASS input.getAttribute("type") for "search" is correctly "search".
PASS input.type for "submit" is correctly "submit".
PASS input.getAttribute("type") for "submit" is correctly "submit".
PASS input.type for "tel" is correctly "tel".
PASS input.getAttribute("type") for "tel" is correctly "tel".
PASS input.type for "telephone" is correctly "text".
PASS input.getAttribute("type") for "telephone" is correctly "telephone".
PASS input.type for "url" is correctly "url".
PASS input.getAttribute("type") for "url" is correctly "url".
PASS input.type for "uri" is correctly "text".
PASS input.getAttribute("type") for "uri" is correctly "uri".
PASS input.type for "" is correctly "text".
PASS input.getAttribute("type") for "" is correctly "".
PASS input.type for "x-unknown" is correctly "text".
PASS input.getAttribute("type") for "x-unknown" is correctly "x-unknown".
PASS input.removeAttribute('type'); input.getAttribute('type') is null
Check dirty flag behavior
PASS input.defaultValue = 'UpdatedDefault'; input.value is "UpdatedDefault"
Changing the type to file should discard the dirty value.
PASS input.value is input.getAttribute('value')
PASS successfullyParsed is true

TEST COMPLETE
